# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank API
  description: API for managing bank cards
  version: 1.0.0
  contact:
    name: Timofey Zaytsev
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /admin/cards:
    post:
      summary: Create a new card
      description: Creates a new card for a user (Admin only)
      operationId: createCard
      tags:
        - Admin Cards
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Card created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCardResponse'
        '400':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Insufficient permissions (ADMIN role required)

    get:
      summary: Get all cards with pagination and filters
      description: Retrieves a paginated list of cards with optional filtering (Admin only)
      operationId: getAllCards
      tags:
        - Admin Cards
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page number (0-indexed)
        - name: size
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 20
          description: Number of items per page
        - name: cardType
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CardType'
          description: Filter by card type
        - name: currency
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Currency'
          description: Filter by currency
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CardStatus'
          description: Filter by card status
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '400':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Insufficient permissions (ADMIN role required)
  /admin/cards/{cardId}:
    get:
      summary: Get card by ID
      description: Retrieves detailed information about a specific card (Admin only)
      operationId: getCardById
      tags:
        - Admin Cards
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier of the card
      responses:
        '200':
          description: Card found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindCardResponse'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Insufficient permissions (ADMIN role required)

    delete:
      summary: Delete card (soft delete)
      description: Changes card status to DELETED or delete from db (Admin only)
      operationId: deleteCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier of the card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteCardRequest'
      responses:
        '204':
          description: Card deleted successfully (no content)
        '400':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Insufficient permissions (ADMIN role required)
  /admin/cards/{cardId}/block:
    put:
      summary: Block a card
      description: Changes card status to BLOCKED (Admin only)
      operationId: blockCard
      tags:
        - Admin Cards
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier of the card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockCardRequest'
      responses:
        '204':
          description: Card blocked successfully (no content)
        '400':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Insufficient permissions (ADMIN role required)
  /admin/cards/{cardId}/activate:
    put:
      summary: Activate a card
      description: Changes card status to ACTIVE (Admin only)
      operationId: activateCard
      tags:
        - Admin Cards
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier of the card
      responses:
        '204':
          description: Card activated successfully (no content)
        '400':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Insufficient permissions (ADMIN role required)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication
  schemas:
...
    CreateCardRequest:
      type: object
      required:
        - userId
        - cardType
        - currency
        - expiryMonth
        - expiryYear
      properties:
        userId:
          type: string
          format: uuid
          description: User ID who will own the card
          example: "123e4567-e89b-12d3-a456-426614174000"
        cardType:
          $ref: '#/components/schemas/CardType'
        currency:
          $ref: '#/components/schemas/Currency'
        expiryMonth:
          type: integer
          description: Expiration month (1-12)
          example: 12
          minimum: 1
          maximum: 12
        expiryYear:
          type: integer
          description: Expiration year (e.g., 2026)
          example: 2026
          minimum: 2024
    CreateCardResponse:
      type: object
      properties:
        cardId:
          type: string
          format: uuid
          description: Generated card ID
          example: "123e4567-e89b-12d3-a456-426614174000"
    FindCardResponse:
      type: object
      properties:
        cardId:
          type: string
          format: uuid
        maskedNumber:
          type: string
        cardType:
          $ref: '#/components/schemas/CardType'
        balance:
          type: number
          format: decimal
          description: Current card balance
        currency:
          $ref: '#/components/schemas/Currency'
        status:
          $ref: '#/components/schemas/CardStatus'
        createdAt:
          type: string
          format: date-time
          description: Card creation timestamp
        expiryDate:
          type: string
          format: date
          description: Card expiration date (YYYY-MM-DD)
    BlockCardRequest:
      type: object
      required:
        - blockReason
      properties:
        blockReason:
          type: string
          description: Reason for blocking the card
          example: "Card reported stolen"
          maxLength: 255
    DeleteCardRequest:
      type: object
      required:
        - softDelete
      properties:
        softDelete:
          type: boolean
          description: Delete soft or not
...
    CardType:
      type: string
      enum:
        - MASTERCARD
        - VISA
        - MIR
      description: Type of the card
    Currency:
      type: string
      enum:
        - USD
        - EUR
        - RUB
      description: Currency code
    CardStatus:
      type: string
      enum:
        - ACTIVE
        - BLOCKED
        - DELETED
        - DEADLINE_HAS_EXPIRED
      description: Current status of the card
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/FindCardResponse'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalPages:
          type: integer
          example: 5
        totalElements:
          type: integer
          example: 100
        last:
          type: boolean
        size:
          type: integer
          example: 20
        number:
          type: integer
          example: 0
        sort:
          $ref: '#/components/schemas/Sort'
        numberOfElements:
          type: integer
          example: 20
        first:
          type: boolean
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        sort:
          $ref: '#/components/schemas/Sort'
        offset:
          type: integer
        pageNumber:
          type: integer
        pageSize:
          type: integer
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    ErrorResponseDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        message:
          type: string
          example: "Bad Request"
        uri:
          type: string
          example: "/api/v1/admin/cards"